Refactor, fix tests for client credentials.
