+++ e2e/tests/oauth/test_oauth_tokens.py
@@ -1,10 +1,10 @@
-from uuid import uuid4
 import pytest
-from time import sleep, time
-from e2e.scripts.config import HELLO_WORLD_API_URL, ID_TOKEN_NHS_LOGIN_PRIVATE_KEY_ABSOLUTE_PATH
-import sys
-from e2e.scripts.config import HELLO_WORLD_API_URL, ID_TOKEN_NHS_LOGIN_PRIVATE_KEY_ABSOLUTE_PATH, MOCK_IDP_BASE_URL
 import requests
+import sys
+from time import sleep, time
+from uuid import uuid4
+
+from e2e.scripts.config import HELLO_WORLD_API_URL, ID_TOKEN_NHS_LOGIN_PRIVATE_KEY_ABSOLUTE_PATH
 
 
 @pytest.mark.asyncio
@@ -377,25 +377,8 @@ class TestTokenExchangeTokens:
 
         assert resp['status_code'] == 400
 
-    @pytest.mark.parametrize("token_expiry_ms, expected_time", [(100000, 100), (500000, 500),(700000, 600), (1000000, 600)])
-    async def test_access_token_override_with_client_credentials(self, token_expiry_ms, expected_time):
-        """
-        Test client credential flow access token can be overridden with a time less than 10 min(600000ms or 600s) 
-        and NOT be overridden with a time greater than 10 min(600000ms or 600s)  
-        """
-        form_data = {
-            "client_assertion_type": "urn:ietf:params:oauth:client-assertion-type:jwt-bearer",
-            "client_assertion": self.oauth.create_jwt('test-1'),
-            "grant_type": 'client_credentials',
-            "_access_token_expiry_ms": token_expiry_ms
-        }
-        
-        resp = await self.oauth.get_token_response(grant_type='client_credentials', data=form_data)
-
-        assert resp['status_code'] == 200
-        assert int(resp['body']['expires_in']) <= expected_time
-    
-    @pytest.mark.parametrize("token_expiry_ms, expected_time", [(100000, 100), (500000, 500),(700000, 600), (1000000, 600)])
+    @pytest.mark.parametrize("token_expiry_ms, expected_time",
+                             [(100000, 100), (500000, 500), (700000, 600), (1000000, 600)])
     async def test_access_token_override_with_token_exchange(self, token_expiry_ms, expected_time):
         """
         Test token exchange flow access token can be overridden with a time less than 10 min(600000ms or 600s)
@@ -411,26 +394,28 @@ class TestTokenExchangeTokens:
             "client_assertion": client_assertion_jwt,
             "_access_token_expiry_ms": token_expiry_ms
         }
-        
+
         resp = await self.oauth.hit_oauth_endpoint("post", "token", data=form_data)
 
         assert resp['status_code'] == 200
         assert int(resp['body']['expires_in']) <= expected_time
-    
-    @pytest.mark.parametrize("token_expiry_ms, expected_time", [(100000, 100), (500000, 500),(700000, 600), (1000000, 600)])
+
+    @pytest.mark.parametrize("token_expiry_ms, expected_time",
+                             [(100000, 100), (500000, 500), (700000, 600), (1000000, 600)])
     async def test_access_token_override_with_authorization_code(self, token_expiry_ms, expected_time):
         """
         Test authorization code flow access token can be overridden with a time less than 10 min(600000ms or 600s)
         and NOT be overridden with a time greater than 10 min(600000ms or 600s)
         """
-        
+
         resp = await self.oauth.get_token_response(grant_type='authorization_code', timeout=token_expiry_ms)
 
         assert resp['status_code'] == 200
         assert int(resp['body']['expires_in']) <= expected_time
-    
+
     @pytest.mark.usefixtures("set_refresh_token")
-    @pytest.mark.parametrize("token_expiry_ms, expected_time", [(100000, 100), (500000, 500),(700000, 600), (1000000, 600)])
+    @pytest.mark.parametrize("token_expiry_ms, expected_time",
+                             [(100000, 100), (500000, 500), (700000, 600), (1000000, 600)])
     async def test_access_token_override_with_refresh_token(self, token_expiry_ms, expected_time):
         """
         Test refresh token flow access token can be overridden with a time less than 10 min(600000ms or 600s)
@@ -444,12 +429,13 @@ class TestTokenExchangeTokens:
             "_refresh_tokens_validity_ms": 599,
             "_access_token_expiry_ms": token_expiry_ms
         }
-        
+
         resp = await self.oauth.get_token_response(grant_type='refresh_token', data=form_data)
-        
+
         assert resp['status_code'] == 200
         assert int(resp['body']['expires_in']) <= expected_time
-    
+
+
 @pytest.mark.asyncio
 class TestTokenRefreshExpiry:
     """Test class to confirm refresh tokens expire after the expected amount of time
